Move utils to libs
* Utils from admin-gui are moved to a lib so it can be reused by another
app.
