+++ apps/admin-gui/src/app/facilities/pages/facility-detail-page/facility-detail-page.component.ts
@@ -3,9 +3,9 @@ import { ActivatedRoute } from '@angular/router';
 import { fadeIn } from '../../../shared/animations/Animations';
 import { SideMenuService } from '../../../core/services/common/side-menu.service';
 import { SideMenuItemService } from '../../../shared/side-menu/side-menu-item.service';
-import { addRecentlyVisited } from '../../../shared/utils';
 import { FacilityService } from '@perun-web-apps/perun/services';
 import { Facility } from '@perun-web-apps/perun/models';
+import { addRecentlyVisited } from '@perun-web-apps/perun/utils';
 
 @Component({
   selector: 'app-facility-detail-page',
