+++ apps/admin-gui/src/app/facilities/pages/facility-select-page/facility-select-page.component.ts
@@ -1,8 +1,8 @@
 import {Component, OnInit} from '@angular/core';
 import {SideMenuService} from '../../../core/services/common/side-menu.service';
-import {getRecentlyVisited, getRecentlyVisitedIds} from '../../../shared/utils';
 import { FacilityService } from '@perun-web-apps/perun/services';
 import { RichFacility } from '@perun-web-apps/perun/models';
+import { getRecentlyVisited, getRecentlyVisitedIds } from '@perun-web-apps/perun/utils';
 
 @Component({
   selector: 'app-facility-select-page',
