+++ apps/admin-gui/src/app/shared/side-menu/side-menu-item.service.ts
@@ -1,9 +1,9 @@
 import {Injectable} from '@angular/core';
 import {TranslateService} from '@ngx-translate/core';
 import {EntityMenuLink, SideMenuItem} from './side-menu.component';
-import {parseFullName} from '../utils';
 import {AuthResolverService} from '../../core/services/common/auth-resolver.service';
 import { Facility, Group, Resource, RichMember, User, Vo } from '@perun-web-apps/perun/models';
+import { parseFullName } from '@perun-web-apps/perun/utils';
 
 @Injectable({
   providedIn: 'root'
