+++ apps/admin-gui/src/app/vos/pages/member-detail-page/member-settings/member-settings-attributes/member-settings-attributes.component.ts
@@ -5,7 +5,6 @@ import {MatDialog} from '@angular/material';
 import {TranslateService} from '@ngx-translate/core';
 import {AttributesListComponent} from '../../../../../shared/components/attributes-list/attributes-list.component';
 import {SelectionModel} from '@angular/cdk/collections';
-import {filterCoreAttributes} from '../../../../../shared/utils';
 import {
   CreateAttributeDialogComponent
 } from '../../../../../shared/components/dialogs/create-attribute-dialog/create-attribute-dialog.component';
@@ -14,6 +13,7 @@ import {
 } from '../../../../../shared/components/dialogs/delete-attribute-dialog/delete-attribute-dialog.component';
 import { AttributesService } from '@perun-web-apps/perun/services';
 import { Attribute } from '@perun-web-apps/perun/models';
+import { filterCoreAttributes } from '@perun-web-apps/perun/utils';
 
 @Component({
   selector: 'app-member-settings-attributes',
